Factor out geocoding code
