+++ app/models/place.rb
@@ -1,7 +1,9 @@
 require 'place/auto_translate'
+require 'place/geocoding'
 
 class Place < ActiveRecord::Base
   include PlaceAutoTranslation
+  include Geocoding
 
   def self.reviewed
     Place.all.map(&:reviewed_version).compact
@@ -52,34 +54,6 @@ class Place < ActiveRecord::Base
 
   def category_ids
     categories.split(',')
-  end
-
-  ## GEOCODING
-  def address
-    "#{street} #{house_number}, #{postal_code} #{city}"
-  end
-
-  def address_changed?
-    street_changed? || city_changed? || house_number_changed? || postal_code_changed?
-  end
-
-  def geocode_with_nodes
-    results = Geocoder.search(address)
-    unless results.any?
-      errors.add(:address, :address_not_found)
-      return false
-    end
-    node_geoms = results.select { |result| result.type == 'node' }
-    other_geoms = results - node_geoms
-    # Prefer point objects (locations, houses, etc.) given by nominatim
-    if node_geoms.any?
-      self.latitude = node_geoms.first.latitude
-      self.longitude = node_geoms.first.longitude
-      # ...else take lat/lon of other geoms (lines, etc.)
-    else
-      self.latitude = other_geoms.first.latitude
-      self.longitude = other_geoms.first.longitude
-    end
   end
 
   ## SANITIZE
